+++ apps/ungdomsytelse-deltaker/src/apps/innsyn/oppgaver/rapporter-inntekt/parts/RapporterInntektOppgave.tsx
@@ -47,6 +47,7 @@ const RapporterInntektOppgave = ({ deltakerNavn, oppgave }: Props) => {
                         oppgaveStatusTekst={getOppgaveStatusText(oppgave)}
                     />
                 </div>
+
                 <Heading level="1" size="large">
                     Rapporter inntekt {månedOgÅr}
                 </Heading>