+++ apps/ungdomsytelse-veileder/src/components/app-header/AppHeader.tsx
@@ -4,13 +4,12 @@ import { InformationSquareIcon, MenuGridIcon, MoonFillIcon, PersonIcon, SunFillI
 import { useThemeContext } from '../../context/ThemeContext';
 import { useVeileder } from '../../context/VeilederContext';
 import { useDrawer } from '../drawer/DrawerContext';
-import { articleList } from '../../pages/info-page/InfoInnhold';
-import ArticleContent from '../../pages/info-page/components/ArticleContent';
+import DrawerArticles from '../../pages/info-page/DrawerArticles';
 
 interface Props {
     visActionsMenu?: boolean;
 }
-const AppHeader = ({ visActionsMenu = true }: Props) => {
+const AppHeader = ({ visActionsMenu = false }: Props) => {
     const { veileder } = useVeileder();
     const { setDarkMode, darkMode } = useThemeContext();
 
@@ -19,7 +18,7 @@ const AppHeader = ({ visActionsMenu = true }: Props) => {
 
     return (
         <InternalHeader>
-            <InternalHeader.Title href="/">Nav Veileder - Ungdomsprogramytelse</InternalHeader.Title>
+            <InternalHeader.Title href="/">Deltakerregistrering - Ungdomsprogramytelsen</InternalHeader.Title>
             <Spacer />
             <InternalHeader.Button
                 aria-label="Bytt mellom lys og mørk modus"
@@ -38,7 +37,7 @@ const AppHeader = ({ visActionsMenu = true }: Props) => {
                     <InternalHeader.Button
                         onClick={(e) => {
                             e.preventDefault();
-                            openDrawer(<ArticleContent articleList={articleList} />, { title: 'Informasjon' });
+                            openDrawer(<DrawerArticles />);
                         }}>
                         <InformationSquareIcon fontSize="1.5rem" title="Informasjonikon" />
                     </InternalHeader.Button>