+++ packages/sif-common-forms-ds/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @navikt/sif-common-forms-ds
 
+## 19.0.8
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+- Updated dependencies
+    - @navikt/sif-common-core-ds@10.2.9
+    - @navikt/sif-common-env@0.1.4
+    - @navikt/sif-common-formik-ds@2.4.7
+    - @navikt/sif-common-ui@0.9.10
+    - @navikt/sif-common-utils@3.49.4
+    - @navikt/sif-validation@1.0.1
+
 ## 19.0.7
 
 ### Patch Changes